Split the address serialization code into modules
